PR updates: move GridDisplay to utilities/index.ts, update comment about cssGridProp, add condition to ensure if the prop is explicitly set to -ms-grid it won't use the cssGrid
