+++ packages/fast-layouts-react/src/column/column.props.ts
@@ -1,5 +1,6 @@
 import React from "react";
-import { GridDisplay, GridGutter } from "../grid/grid.props";
+import { GridDisplay } from "../utilities";
+import { GridGutter } from "../grid/grid.props";
 import { ColumnClassNamesContract } from "./column";
 import { ManagedClasses } from "@microsoft/fast-jss-manager-react";
 
@@ -33,7 +34,7 @@ export interface ColumnHandledProps extends ColumnManagedClasses {
     gutter?: GridGutter;
 
     /**
-     * The CSS grid display, "grid" when true and "-ms-grid" when false
+     * The value assigned to the CSS display property
      * Provide this prop when doing server side rendering
      */
     cssGridPropertyName?: GridDisplay;
