+++ packages/fast-layouts-react/src/grid/index.ts
@@ -4,7 +4,6 @@ import {
     Grid as BaseGrid,
     GridAlignment,
     GridClassNamesContract,
-    GridDisplay,
     GridGutter,
     GridHandledProps as BaseGridHandledProps,
     GridManagedClasses,
@@ -30,5 +29,4 @@ export {
     GridTag,
     GridAlignment,
     GridGutter,
-    GridDisplay,
 };
