+++ packages/fast-layouts-react/src/utilities/index.ts
@@ -4,3 +4,8 @@ export * from "./breakpoint-tracker";
 
 export * from "./breakpoints";
 export * from "./string-extensions";
+
+/**
+ * The CSS grid display style
+ */
+export type GridDisplay = "grid" | "-ms-grid";
